+++ udata/core/user/commands.py
@@ -20,14 +20,19 @@ def grp():
 
 
 @grp.command()
-def create():
+@click.option("--first-name")
+@click.option("--last-name")
+@click.option("--email")
+@click.option("--password")
+@click.option("--admin", is_flag=True)
+def create(first_name, last_name, email, password, admin):
     """Create a new user"""
     data = {
-        "first_name": click.prompt("First name"),
-        "last_name": click.prompt("Last name"),
-        "email": click.prompt("Email"),
-        "password": click.prompt("Password", hide_input=True),
-        "password_confirm": click.prompt("Confirm Password", hide_input=True),
+        "first_name": first_name or click.prompt("First name"),
+        "last_name": last_name or click.prompt("Last name"),
+        "email": email or click.prompt("Email"),
+        "password": password or click.prompt("Password", hide_input=True),
+        "password_confirm": password or click.prompt("Confirm Password", hide_input=True),
     }
     # Until https://github.com/mattupstate/flask-security/issues/672 is fixed
     with current_app.test_request_context():
@@ -37,6 +42,9 @@ def create():
         del data["password_confirm"]
         data["confirmed_at"] = datetime.utcnow()
         user = datastore.create_user(**data)
+        if admin:
+            role = datastore.find_or_create_role("admin")
+            datastore.add_role_to_user(user, role)
         success("User(id={u.id} email={u.email}) created".format(u=user))
         return user
     errors = "\n".join("\n".join([str(m) for m in e]) for e in form.errors.values())