+++ CHANGELOG.md
@@ -8,7 +8,8 @@
   - Change the redirects URLs in ProConnect:
     - From https://www.data.gouv.fr/proconnect/auth to https://www.data.gouv.fr/api/1/proconnect/auth
     - From https://www.data.gouv.fr/proconnect/logout to https://www.data.gouv.fr/api/1/proconnect/logout
-
+- **breaking change** feat: topic elements [#3318](https://github.com/opendatateam/udata/pull/3318) [#3416](https://github.com/opendatateam/udata/pull/3416) [#3417](https://github.com/opendatateam/udata/pull/3417) [#3418](https://github.com/opendatateam/udata/pull/3418)
+- Allow to send arguments to the `udata user create` command for CI [#3419](https://github.com/opendatateam/udata/pull/3419)
 
 ## 10.9.0 (2025-08-28)
 
@@ -97,7 +98,6 @@
 - Allow global admins to manually run harvesters in production [#3340](https://github.com/opendatateam/udata/pull/3340)
 - Add monthly metrics to datasets, dataservices, reuses and organizations [#3342](https://github.com/opendatateam/udata/pull/3342)
 - Fix user avatar deletion [#3346](https://github.com/opendatateam/udata/pull/3346)
-- **breaking change** feat: topic elements [#3318](https://github.com/opendatateam/udata/pull/3318)
 
 ## 10.4.2 (2025-06-05)